refactor(evento_detalhes): extract EscalaTextoBuilder, GerenciarSetlistDialog, EventoSetlistTab, EventoEscalaTab
